+++ src/functions/configuration.js
@@ -82,6 +82,7 @@ async function checkConfigFile(file, moduleName) {
         let newConfig = exampleFile.configElements ? [] : {};
         if (exampleFile.elementLimits) configData = require('./scnx-integration').verifyLimitedConfigElementFile(client, exampleFile, configData);
 
+        let skipOverwrite = false;
         if (exampleFile.skipContentCheck) newConfig = configData;
         else if (exampleFile.configElements) {
             if (!Array.isArray(configData)) {
@@ -105,7 +106,10 @@ async function checkConfigFile(file, moduleName) {
             }
         } else {
             for (const field of exampleFile.content) {
-                if (exampleFile.content.find(f => f.elementToggle) && !configData[exampleFile.content.find(f => f.elementToggle)]) continue;
+                if (exampleFile.content.find(f => f.elementToggle) && !configData[exampleFile.content.find(f => f.elementToggle).name]) {
+                    skipOverwrite = true;
+                    continue;
+                }
                 const dependsOnField = field.dependsOn ? exampleFile.content.find(f => f.name === field.dependsOn) : null;
                 if (field.dependsOn && !dependsOnField) return reject(`Depends-On-Field ${field.dependsOn} does not exist.`);
                 if (dependsOnField && !(typeof configData[dependsOnField.name] === 'undefined' ? (dependsOnField.default[client.locale] || dependsOnField.default['en']) : configData[dependsOnField.name])) {
@@ -173,7 +177,7 @@ async function checkConfigFile(file, moduleName) {
             });
         }
 
-        if (!isEqual(configData, newConfig) || forceOverwrite) {
+        if (!skipOverwrite && (!isEqual(configData, newConfig) || forceOverwrite)) {
             if (!fs.existsSync(`${client.configDir}/${moduleName}`) && moduleName) fs.mkdirSync(`${client.configDir}/${moduleName}`);
             jsonfile.writeFileSync(`${client.configDir}${builtIn ? '' : '/' + moduleName}/${exampleFile.filename}`, newConfig, {spaces: 2});
             logger.info(localize('config', 'saved-file', {
@@ -195,16 +199,16 @@ async function checkConfigFile(file, moduleName) {
  */
 async function checkModuleConfig(moduleName, afterCheckEventFile = null) {
     return new Promise(async (resolve, reject) => {
-            const moduleConf = require(`../../modules/${moduleName}/module.json`);
-            if ((moduleConf['config-example-files'] || []).length === 0) return resolve();
-            try {
-                for (const v of moduleConf['config-example-files']) await checkConfigFile(v, moduleName);
-                resolve();
-            } catch (r) {
-                reject(r);
-            }
-            if (afterCheckEventFile) require(`../../modules/${moduleName}/${afterCheckEventFile}`).afterCheckEvent(config);
+        const moduleConf = require(`../../modules/${moduleName}/module.json`);
+        if ((moduleConf['config-example-files'] || []).length === 0) return resolve();
+        try {
+            for (const v of moduleConf['config-example-files']) await checkConfigFile(v, moduleName);
+            resolve();
+        } catch (r) {
+            reject(r);
         }
+        if (afterCheckEventFile) require(`../../modules/${moduleName}/${afterCheckEventFile}`).afterCheckEvent(config);
+    }
     );
 }
